Protocol buffers for factor graph serialization.

+++ src/main/scala/org/deepdive/inference/models/Factor.scala
@@ -5,7 +5,7 @@ case class Factor(id: Long, factorFunction: String, weightId: Long,
   def toCSVRow = Array(id.toString, weightId.toString, factorFunction.toString)
 }
 
-case class FactorVariable(factorId: Long, position: Integer, positive: Boolean, 
+case class FactorVariable(factorId: Long, position: Long, positive: Boolean, 
   variableId: Long) extends  CSVFormattable {
   def toCSVRow =  Array(factorId.toString, variableId.toString, position.toString, positive.toString)
 }
